requirements.txt
flask
mongoengine
python-dotenv
pyjwt
flask-cors
openai
# pillow
# easyocr 
# opencv-python 
# numpy
